tyre_sentiment_cli/requirements.txt
typer==0.16.0
openai==1.51.2
rich==13.9.4
python-dotenv==1.0.1
feedparser==6.0.11
httpx==0.27.2
requests==2.32.3
